add install option to install fonts

+++ install.sh
@@ -1,5 +1,30 @@
 #!/bin/bash
 
+print_help () {
+    echo "Usage: $0 [-f] [-h]"
+    echo ""
+    echo "  -f    install additional font (will install Ubuntu Mono Nerd Font from https://github.com/ryanoasis/nerd-fonts)"
+    echo "  -h    show help message"
+}
+
+install_fonts=false # do not install fonts by default
+
+while getopts "fh" opt; do
+    case ${opt} in
+        f)
+            install_fonts=true
+            ;;
+        h)
+            print_help
+            exit
+            ;;
+        \?)
+            print_help
+            exit 1
+            ;;
+    esac
+done
+
 sudo apt-get update
 
 # install dependencies
@@ -23,24 +48,33 @@ cp zshrc $HOME/.zshrc
 # source for P9KGT: https://github.com/lucatrv/dotfiles
 cp zshrc_P9KGT $HOME/.zshrc_P9KGT
 
-# install Ubuntu Mono Nerd Font from https://github.com/ryanoasis/nerd-fonts
-unzip UbuntuMono.zip -d UbuntuMono
+# install additional fonts
+if [ $install_fonts = true ]; then
+    echo "Start install Ubuntu Mono Nerd Font..."
 
-# move fonts to $HOME/.font directory
-if [ ! -f "$HOME/.fonts" ]; then
-    mkdir -p $HOME/.fonts
+    # install Ubuntu Mono Nerd Font from https://github.com/ryanoasis/nerd-fonts
+    unzip UbuntuMono.zip -d UbuntuMono
+
+    # move fonts to $HOME/.font directory
+    if [ ! -f "$HOME/.fonts" ]; then
+        mkdir -p $HOME/.fonts
+    fi
+
+    cp UbuntuMono/* $HOME/.fonts/
+
+    # rebuild the font cache
+    fc-cache -f -v
+
+    # delete nerd-fonts directory after installing the font
+    rm -rf UbuntuMono
+    echo "install done"
 fi
-
-cp UbuntuMono/* $HOME/.fonts/
-
-# rebuild the font cache
-fc-cache -f -v
-
-# delete nerd-fonts directory after installing the font
-rm -rf UbuntuMono
 
 # change default shell to zsh
 chsh -s /bin/zsh
 
 echo "Remember login again to reload default shell settings"
-echo "And change terminal font to \"UbuntuMono Nerd Font Regular\""
+
+if [ $install_fonts = true ]; then
+    echo "And change terminal font to \"UbuntuMono Nerd Font Regular\""
+fi
